+++ react/__mocks__/vtex.payment-flags.tsx
@@ -1,0 +1,8 @@
+import React, { Fragment } from 'react'
+
+export const utils = {
+  translatePaymentGroup: () => '',
+  displayPayment: () => '',
+}
+
+export const PaymentFlag = () => <Fragment />